setup zsh for mac

+++ .zshrc
@@ -10,22 +10,24 @@ plugins=(git autojump)
 # oh-my-zsh
 source $ZSH/oh-my-zsh.sh
 
-# aliases
-alias up="sudo apt update && sudo apt upgrade && sudo apt dist-upgrade && sudo apt autoremove"
+# env vars and aliases by env
+if [[ "$(uname -s)" == "Darwin" ]]; then
+  export ZPLUG_HOME=$(brew --prefix)/opt/zplug
+  alias up="brew update & brew upgrade"
+elif [[ "$(uname -s)" == "Linux" ]]; then
+  alias up="sudo apt update && sudo apt upgrade && sudo apt dist-upgrade && sudo apt autoremove"
+fi
 
-# zinit installer
-source "$HOME/.zinit/bin/zinit.zsh"
-autoload -Uz _zinit
-(( ${+_comps} )) && _comps[zinit]=_zinit
+# zplug
+source $ZPLUG_HOME/init.zsh
 
-# zinit plugins
-zinit light zdharma/fast-syntax-highlighting
-zinit light zsh-users/zsh-autosuggestions
-zinit light zsh-users/zsh-completions
+# zplug plugins
+zplug "zsh-users/zsh-syntax-highlighting"
+zplug "zsh-users/zsh-autosuggestions"
+zplug "zsh-users/zsh-completions"
+zplug load
 
 # nvm
-export NVM_DIR="$([ -z "${XDG_CONFIG_HOME-}" ] && printf %s "${HOME}/.nvm" || printf %s "${XDG_CONFIG_HOME}/nvm")"
+export NVM_DIR="$HOME/.nvm"
 [ -s "$NVM_DIR/nvm.sh" ] && \. "$NVM_DIR/nvm.sh"
-
-# go
-export PATH=$PATH:/usr/local/go/bin
+[ -s "$NVM_DIR/bash_completion" ] && \. "$NVM_DIR/bash_completion"
